pyproject.toml
[project]
name = "prompt-panda"
version = "0.1.0"
description = "Secure local AI assistant powered by Ollama"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
web = ["fastapi>=0.111", "uvicorn[standard]>=0.30"]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
all = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-telegram-bot>=21.0",
    "discord.py>=2.3",
]

[project.scripts]
prompt-panda = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
